refactor : separate filtering option select to another file
